Removed extra logs ,modified ci.yaml
