--- sideInput_function/go.mod
@@ -1,15 +1,0 @@
-module github.com/numaproj/numaflow-sinks/argoworkflow
-
-go 1.21
-
-require github.com/numaproj/numaflow-go v0.5.2
-
-require (
-	github.com/golang/protobuf v1.5.3 // indirect
-	golang.org/x/net v0.9.0 // indirect
-	golang.org/x/sys v0.7.0 // indirect
-	golang.org/x/text v0.9.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20230525234030-28d5490b6b19 // indirect
-	google.golang.org/grpc v1.57.0 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
-)
